phase-63: HS-63-03 the feature glue out of WebRuntime — wake, devices, dictation capture
WebRuntime(DictationCaptureMixin, WakeWordGlueMixin, DeviceGlueMixin):
dictation_capture.py (413: transcribe-and-type, hotkey, tmux agent
reply, voice-command dispatch), wake_glue.py (the HS-60 section
verbatim), device_glue.py (the AIPI-Lite handlers). web_runtime.py is
1,763 lines (was 2,635). The verbatim proof is one line: the only
original body line lost is 'class WebRuntime:'.

Test edits are exactly the censused patch-target paths (5 sites, 3
files; assertions byte-identical) — with an honest find: the first wake
retarget pointed at wake_glue and two tests passed COINCIDENTALLY (the
real global lookup happens in dictation_capture via
_maybe_run_dictation_pipeline; the unpatched real pipeline ran
silently); the third test failed loudly and exposed it. All sites now
point where the lookup happens, and the misleading unused imports were
trimmed from wake_glue/device_glue.

Full suite 2768 passed, 17 skipped.
