+++ holdspeak/runtime/__init__.py
@@ -1,0 +1,6 @@
+"""WebRuntime's concern modules (Phase 63).
+
+Mixin classes composed by `holdspeak.web_runtime.WebRuntime` — the backend
+twin of the Phase-54 frontend partials. Bodies are verbatim moves; `self`
+is the runtime.
+"""
