+++ holdspeak/web_runtime.py
@@ -109,7 +109,16 @@ def _dictation_journal_repo():
         return None
 
 
-class WebRuntime:
+from .runtime.device_glue import DeviceGlueMixin
+from .runtime.dictation_capture import DictationCaptureMixin
+from .runtime.wake_glue import WakeWordGlueMixin
+
+
+class WebRuntime(
+    DictationCaptureMixin,
+    WakeWordGlueMixin,
+    DeviceGlueMixin,
+):
     """Web-first runtime: owns the web server, hotkey/device capture, the
     meeting session, and the MIR plugin pipeline.
 
@@ -1584,887 +1593,6 @@ class WebRuntime:
         if session is None:
             return None
         return session.edit_action_item(item_id, task=task, owner=owner, due=due)
-
-    def _transcribe_and_type(
-        self,
-        audio: np.ndarray,
-        *,
-        on_complete: Optional[Callable[[str], None]] = None,
-        agent_reply_session: Any | None = None,
-    ) -> None:
-        """Run transcription, text processing, and typing for a captured chunk.
-
-        Shared between the local hotkey path and the device-driven
-        voice-typing path (HS-14-05). Always flips voice state back
-        to ``idle`` in its ``finally``. ``on_complete`` (HS-14-07)
-        receives the typed text on success and is intentionally
-        invoked outside the typing try-block — typing failures
-        still surface the transcript to the device.
-        """
-        completed_text: Optional[str] = None
-        with self.transcription_lock:
-            try:
-                text = self._ensure_transcriber_loaded().transcribe(audio)
-                if not text:
-                    self._set_runtime_activity(
-                        "complete",
-                        source="dictation",
-                        label="No speech",
-                        detail="No speech detected.",
-                        last_event="dictation_no_speech",
-                        last_error="",
-                    )
-                    return
-                text = self.text_processor.process(text)
-                # HS-52-04: voice command dispatch. A configured, enabled keyword fires
-                # an action instead of being typed; on a match we return early and type
-                # nothing. Off by default and on no match this is inert (byte-identical).
-                voice_command = self._maybe_dispatch_voice_command(text, agent_reply_session)
-                if voice_command is not None:
-                    if voice_command.ok:
-                        self._set_runtime_activity(
-                            "complete",
-                            source="dictation",
-                            label="Command",
-                            detail=voice_command.preview,
-                            last_event="voice_command_fired",
-                            last_error="",
-                        )
-                        self._mark_first_dictation()
-                    else:
-                        with self.state_lock:
-                            self.runtime_status["last_error"] = (
-                                f"Voice command failed: {voice_command.error}"
-                            )
-                        self._set_runtime_activity(
-                            "error",
-                            source="dictation",
-                            label="Command failed",
-                            detail=voice_command.preview,
-                            last_event="voice_command_failed",
-                            last_error=voice_command.error,
-                        )
-                    return
-                self._set_runtime_activity(
-                    "processing",
-                    source="dictation",
-                    detail="Processing dictation.",
-                    last_event="dictation_processing",
-                    last_error="",
-                )
-                text = self._maybe_run_dictation_pipeline(
-                    text,
-                    audio_duration_s=len(audio) / 16000.0,
-                    transcribed_at=datetime.now(),
-                    agent_reply_session=agent_reply_session,
-                )
-                completed_text = text
-                with self.state_lock:
-                    self.runtime_status["last_transcription"] = text
-                    self.runtime_status["last_error"] = ""
-                print(f"-> {text}")
-                delivered = self._try_tmux_agent_reply(text, agent_reply_session)
-                if delivered:
-                    self._set_runtime_activity(
-                        "complete",
-                        source="dictation",
-                        label="Sent",
-                        detail="Sent dictated text to the agent session.",
-                        last_event="dictation_delivered",
-                        last_error="",
-                    )
-                    self._mark_first_dictation()
-                if not delivered and self.typer is not None:
-                    try:
-                        paste_target_profile = self._paste_target_profile(agent_reply_session)
-                        self._set_runtime_activity(
-                            "typing",
-                            source="dictation",
-                            detail="Typing dictated text.",
-                            last_event="dictation_typing",
-                            last_error="",
-                        )
-                        self.typer.type_text(
-                            text,
-                            target_profile=paste_target_profile,
-                            submit=agent_reply_session is not None,
-                        )
-                        self._set_runtime_activity(
-                            "complete",
-                            source="dictation",
-                            label="Typed",
-                            detail="Dictated text was inserted.",
-                            last_event="dictation_typed",
-                            last_error="",
-                        )
-                        self._mark_first_dictation()
-                    except Exception as exc:
-                        with self.state_lock:
-                            self.runtime_status["last_error"] = f"Typing failed: {exc}"
-                            self.runtime_status["text_injection_enabled"] = False
-                            self.runtime_status["text_injection_error"] = f"{type(exc).__name__}: {exc}"
-                        self._set_runtime_activity(
-                            "error",
-                            source="dictation",
-                            detail="Typing failed.",
-                            last_event="dictation_typing_failed",
-                            last_error=f"{type(exc).__name__}: {exc}",
-                        )
-                        log.warning(f"Typing failed in web mode: {exc}")
-            except Exception as exc:
-                with self.state_lock:
-                    self.runtime_status["last_error"] = f"Transcription failed: {exc}"
-                self._set_runtime_activity(
-                    "error",
-                    source="dictation",
-                    detail="Transcription failed.",
-                    last_event="dictation_transcription_failed",
-                    last_error=f"{type(exc).__name__}: {exc}",
-                )
-                log.error(f"Transcription failed in web mode: {exc}")
-            finally:
-                self._set_voice_state("idle", update_activity=False)
-        if on_complete is not None and completed_text is not None:
-            try:
-                on_complete(completed_text)
-            except Exception as exc:
-                log.warning(f"on_complete hook raised: {exc}")
-
-    def _kick_off_transcribe(
-        self,
-        audio: np.ndarray,
-        *,
-        on_complete: Optional[Callable[[str], None]] = None,
-        agent_reply_session: Any | None = None,
-        source: str = "dictation",
-    ) -> None:
-        if len(audio) < 1600:
-            self._set_voice_state("idle", update_activity=False)
-            self._set_runtime_activity(
-                "complete",
-                source=source,
-                label="Too short",
-                detail="Recording was too short.",
-                last_event="dictation_too_short",
-                last_error="",
-            )
-            return
-        self._set_voice_state(
-            "transcribing",
-            source=source,
-            detail="Transcribing audio.",
-            last_event="dictation_transcribing",
-            last_error="",
-        )
-        threading.Thread(
-            target=lambda: self._transcribe_and_type(
-                audio,
-                on_complete=on_complete,
-                agent_reply_session=agent_reply_session,
-            ),
-            daemon=True,
-        ).start()
-
-    def _maybe_dispatch_voice_command(
-        self, text: str, agent_reply_session: Any | None = None
-    ) -> Any:
-        # HS-52-04: thin delegate to the carved dispatch seam. Injects the runtime
-        # typer for `type_text` macros and surfaces a matched command as a runtime
-        # activity. Returns a VoiceCommandResult if a command fired (caller types
-        # nothing), else None.
-        def _type(t: str) -> None:
-            if self.typer is not None:
-                self.typer.type_text(
-                    t, target_profile=self._paste_target_profile(agent_reply_session)
-                )
-
-        def _activity(label: str) -> None:
-            self._set_runtime_activity(
-                "processing",
-                source="dictation",
-                label=label,
-                detail=label,
-                last_event="voice_command_match",
-                last_error="",
-            )
-
-        return dispatch_voice_command(
-            text,
-            config=self.config,
-            type_writer=_type,
-            on_activity=_activity,
-        )
-
-    def _maybe_run_dictation_pipeline(
-        self,
-        text: str,
-        *,
-        audio_duration_s: float,
-        transcribed_at: datetime,
-        agent_reply_session: Any | None = None,
-        journal_source: str = "dictation",
-    ) -> str:
-        # HS-52-01: the orchestration was carved out of this god-object into
-        # `holdspeak.dictation_runner`; this stays as the thin delegate the
-        # transcription path calls. Behaviour is unchanged.
-        return run_dictation_pipeline(
-            text,
-            config=self.config,
-            server=self.server,
-            audio_duration_s=audio_duration_s,
-            transcribed_at=transcribed_at,
-            agent_reply_session=agent_reply_session,
-            journal_source=journal_source,
-        )
-
-    def _paste_target_profile(self, agent_reply_session: Any | None) -> str | None:
-        if agent_reply_session is None:
-            return None
-        try:
-            from holdspeak.agent_device import target_profile_override_for_agent
-
-            return target_profile_override_for_agent(agent_reply_session)
-        except Exception:
-            return None
-
-    def _try_tmux_agent_reply(self, text: str, agent_reply_session: Any | None) -> bool:
-        pane = self._agent_tmux_pane(agent_reply_session)
-        if not pane:
-            return False
-        try:
-            from holdspeak.tmux_transport import send_text_to_pane
-
-            send_text_to_pane(pane=pane, text=text, submit=True)
-            return True
-        except Exception as exc:
-            with self.state_lock:
-                self.runtime_status["last_error"] = f"tmux reply failed; fell back to typing: {exc}"
-            log.warning(f"tmux reply failed; falling back to text injection: {exc}")
-            return False
-
-    def _agent_tmux_pane(self, agent_reply_session: Any | None) -> str | None:
-        if agent_reply_session is None:
-            return None
-        pane = getattr(agent_reply_session, "tmux_pane", None)
-        return str(pane).strip() if pane else None
-
-    def _agent_reply_deliverable(self, agent_reply_session: Any | None) -> bool:
-        if agent_reply_session is None:
-            return True
-        if self._agent_tmux_pane(agent_reply_session):
-            return True
-        return self.typer is not None
-
-    # ── HS-60: the wake word ────────────────────────────────────────────
-
-    def _sync_wake_word(self) -> None:
-        """Start/stop the wake listener to match config (live via settings)."""
-        want = bool(getattr(getattr(self.config, "wake_word", None), "enabled", False))
-        have = self._wake_listener is not None
-        if want and not have:
-            self._start_wake_listener()
-        elif not want and have:
-            self._stop_wake_listener()
-
-    def _start_wake_listener(self) -> None:
-        from .wake_word import (
-            FRAME_SAMPLES,
-            SAMPLE_RATE as WAKE_RATE,
-            OpenWakeWordDetector,
-            WakeWordListener,
-            wake_word_available,
-        )
-
-        if not wake_word_available():
-            log.warning(
-                "Wake word is enabled but the engine is not installed; "
-                "install it with: pip install 'holdspeak[wakeword]'"
-            )
-            return
-        cfg = self.config.wake_word
-        try:
-            detector = OpenWakeWordDetector(cfg.model)
-        except Exception:
-            # First enable: fetch the models — the feature's ONE network
-            # moment (~7 MB from the openWakeWord GitHub releases), stated in
-            # the settings copy and the docs.
-            try:
-                from .wake_word import download_wake_models
-
-                log.info(
-                    f"Downloading the wake models for {cfg.model!r} "
-                    "(one-time, from the openWakeWord GitHub releases)…"
-                )
-                download_wake_models(cfg.model)
-                detector = OpenWakeWordDetector(cfg.model)
-            except Exception as exc:
-                log.warning(f"Wake model {cfg.model!r} unavailable: {exc}")
-                return
-        import queue as queue_mod
-
-        try:
-            import sounddevice as sd
-        except Exception as exc:  # pragma: no cover - portaudio missing
-            log.warning(f"Wake word needs sounddevice: {exc}")
-            return
-
-        wake_queue: Any = queue_mod.Queue(maxsize=64)
-
-        def _cb(indata, _frames, _time, _status) -> None:
-            try:
-                wake_queue.put_nowait(indata[:, 0].copy())
-            except queue_mod.Full:  # drop, never block the audio thread
-                pass
-
-        try:
-            stream = sd.InputStream(
-                samplerate=WAKE_RATE,
-                channels=1,
-                dtype="int16",
-                blocksize=FRAME_SAMPLES,
-                callback=_cb,
-            )
-            stream.start()
-        except Exception as exc:
-            log.warning(f"Wake word could not open the microphone: {exc}")
-            return
-        self._wake_queue = wake_queue
-        self._wake_stream = stream
-
-        def _frames():
-            # Self-healing floor respect: while ANY owner holds the audio
-            # floor (hotkey, device, meeting, or a wake capture), the
-            # listener pauses (frames drain unscored); it resumes when free.
-            listener = self._wake_listener
-            if listener is not None:
-                if self.voice_session.active_owner is not None:
-                    listener.pause()
-                else:
-                    listener.resume()
-            try:
-                return wake_queue.get(timeout=0.5)
-            except queue_mod.Empty:
-                # Keep the loop alive through quiet stream hiccups.
-                return np.zeros(FRAME_SAMPLES, dtype=np.int16)
-
-        self._wake_listener = WakeWordListener(
-            detector=detector,
-            frames=_frames,
-            on_detect=self._on_wake_detect,
-            threshold=cfg.threshold,
-        )
-        self._wake_listener.start()
-        log.info(f"Wake word active: {cfg.model!r} (threshold {cfg.threshold})")
-
-    def _stop_wake_listener(self) -> None:
-        listener, self._wake_listener = self._wake_listener, None
-        if listener is not None:
-            try:
-                listener.stop()
-            except Exception:
-                pass
-        stream, self._wake_stream = self._wake_stream, None
-        if stream is not None:
-            try:
-                stream.stop()
-                stream.close()
-            except Exception:
-                pass
-        self._wake_queue = None
-
-    def _on_wake_detect(self, score: float) -> None:
-        """A detection: acquire the floor, arm visibly, capture, hand off.
-
-        Runs on the listener thread; the frame queue keeps filling from the
-        stream callback, so the capture reads the same source.
-        """
-        from .wake_word import ArmedCapture, FRAME_SAMPLES
-
-        cfg = self.config.wake_word
-        if not self.voice_session.acquire("wake"):
-            return  # someone holds the floor; never contend
-        audio = None
-        try:
-            self._set_runtime_activity(
-                "armed",
-                source="wake",
-                label="Armed",
-                detail=f"Say your sentence ({int(cfg.armed_window_seconds)} s window).",
-                last_event="wake_armed",
-                last_error="",
-            )
-            try:
-                self.server.broadcast(
-                    "wake_armed",
-                    {"window_seconds": cfg.armed_window_seconds, "score": round(float(score), 3)},
-                )
-            except Exception:
-                pass
-            capture = ArmedCapture(window_seconds=cfg.armed_window_seconds)
-            # A hard iteration cap so a dead stream can never wedge the floor.
-            max_iterations = int((cfg.armed_window_seconds + 20.0) / 0.08)
-            queue_ref = self._wake_queue
-            for _ in range(max_iterations):
-                if capture.state in ("captured", "expired"):
-                    break
-                if self.runtime_stop_event.is_set() or queue_ref is None:
-                    break
-                try:
-                    frame = queue_ref.get(timeout=1.0)
-                except Exception:
-                    continue
-                capture.feed(frame)
-            audio = capture.result()
-        finally:
-            self.voice_session.release("wake")
-        if audio is None or len(audio) < 1600:
-            self._set_runtime_activity(
-                "complete",
-                source="wake",
-                label="Disarmed",
-                detail="Nothing was spoken.",
-                last_event="wake_disarmed",
-                last_error="",
-            )
-            return
-        self._transcribe_wake(audio)
-
-    def _transcribe_wake(self, audio: np.ndarray) -> None:
-        """The wake outcome: the NORMAL pipeline, then preview or (opt-in) type.
-
-        `action="preview"` (the default) journals the run (source `wake`),
-        stores a one-shot preview token, and broadcasts `wake_preview` —
-        it NEVER types. `action="type"` is the user's explicit opt-in and
-        behaves like a hotkey run's tail.
-        """
-        cfg = self.config.wake_word
-        with self.transcription_lock:
-            try:
-                self._set_runtime_activity(
-                    "transcribing",
-                    source="wake",
-                    detail="Turning your speech into text…",
-                    last_event="wake_transcribing",
-                    last_error="",
-                )
-                text = self._ensure_transcriber_loaded().transcribe(audio)
-                if not text:
-                    self._set_runtime_activity(
-                        "complete",
-                        source="wake",
-                        label="No speech",
-                        detail="No speech detected.",
-                        last_event="wake_no_speech",
-                        last_error="",
-                    )
-                    return
-                text = self.text_processor.process(text)
-                final = self._maybe_run_dictation_pipeline(
-                    text,
-                    audio_duration_s=len(audio) / 16000.0,
-                    transcribed_at=datetime.now(),
-                    journal_source="wake",
-                )
-                if cfg.action == "type":
-                    self._set_runtime_activity(
-                        "typing",
-                        source="wake",
-                        detail="Typing into the active app.",
-                        last_event="wake_typing",
-                        last_error="",
-                    )
-                    self.typer.type_text(final)
-                    self._set_runtime_activity(
-                        "complete",
-                        source="wake",
-                        label="Typed",
-                        detail=final[:120],
-                        last_event="wake_typed",
-                        last_error="",
-                    )
-                    return
-                # The preview default: one active preview at a time.
-                import uuid as uuid_mod
-
-                token = uuid_mod.uuid4().hex
-                self.wake_previews.clear()
-                self.wake_previews[token] = {
-                    "text": final,
-                    "transcript": text,
-                    "created_at": datetime.now().isoformat(),
-                }
-                try:
-                    self.server.broadcast(
-                        "wake_preview",
-                        {"token": token, "transcript": text, "text": final},
-                    )
-                except Exception:
-                    pass
-                self._set_runtime_activity(
-                    "complete",
-                    source="wake",
-                    label="Preview ready",
-                    detail=final[:120],
-                    last_event="wake_preview",
-                    last_error="",
-                )
-            except Exception as exc:
-                self._set_runtime_activity(
-                    "error",
-                    source="wake",
-                    detail="Wake transcription failed.",
-                    last_event="wake_failed",
-                    last_error=f"{type(exc).__name__}: {exc}",
-                )
-
-    def consume_wake_preview(self, token: str) -> Optional[str]:
-        """One-shot: return the stored preview text and burn the token."""
-        entry = self.wake_previews.pop(str(token or ""), None)
-        return None if entry is None else str(entry.get("text", ""))
-
-    def _type_wake_preview(self, token: str) -> Optional[str]:
-        """The Type-it route's handler: burn the token, type the stored text."""
-        text = self.consume_wake_preview(token)
-        if text is None:
-            return None
-        try:
-            self.typer.type_text(text)
-        except Exception as exc:
-            self._set_runtime_activity(
-                "error",
-                source="wake",
-                detail="Typing the wake preview failed.",
-                last_event="wake_type_failed",
-                last_error=f"{type(exc).__name__}: {exc}",
-            )
-            return None
-        self._set_runtime_activity(
-            "complete",
-            source="wake",
-            label="Typed",
-            detail=text[:120],
-            last_event="wake_preview_typed",
-            last_error="",
-        )
-        return text
-
-    def _on_hotkey_press(self) -> None:
-        if self.runtime_stop_event.is_set():
-            return
-        if self.recorder is None:
-            self._set_runtime_activity(
-                "error",
-                source="hotkey",
-                detail="Voice typing hotkey is unavailable.",
-                last_event="dictation_hotkey_unavailable",
-                last_error=str(self.runtime_status.get("global_hotkey_error") or ""),
-            )
-            return
-        # HS-32-03: no explicit "is a meeting active?" check — the shared
-        # `voice_session` arbiter is the single owner model. While a meeting
-        # holds the floor (owner="meeting"), `begin()` returns False here.
-        try:
-            accepted = self.voice_session.begin(self.recorder, owner="hotkey")
-        except Exception as exc:
-            with self.state_lock:
-                self.runtime_status["last_error"] = f"Recording failed: {exc}"
-            self._set_voice_state(
-                "idle",
-                source="hotkey",
-                detail="Recording failed.",
-                last_event="dictation_recording_failed",
-                last_error=f"{type(exc).__name__}: {exc}",
-            )
-            log.error(f"Recording failed in web mode: {exc}")
-            return
-        if not accepted:
-            log.info("hotkey_press_ignored_session_active")
-            self._set_runtime_activity(
-                "complete",
-                source="hotkey",
-                label="Busy",
-                detail="Another HoldSpeak audio session is active.",
-                last_event="dictation_recording_busy",
-                last_error="",
-            )
-            return
-        self._set_voice_state(
-            "recording",
-            source="hotkey",
-            detail="HoldSpeak is listening.",
-            last_event="dictation_recording_started",
-            last_error="",
-        )
-
-    def _on_hotkey_release(self) -> None:
-        # No meeting check: `end("hotkey")` returns None when the hotkey
-        # doesn't own the floor (e.g. a meeting holds it), so this is a no-op.
-        try:
-            audio = self.voice_session.end(owner="hotkey")
-        except Exception as exc:
-            with self.state_lock:
-                self.runtime_status["last_error"] = f"Recording error: {exc}"
-            self._set_voice_state(
-                "idle",
-                source="hotkey",
-                detail="Recording stop failed.",
-                last_event="dictation_recording_stop_failed",
-                last_error=f"{type(exc).__name__}: {exc}",
-            )
-            log.error(f"Recording error in web mode: {exc}")
-            return
-        if audio is None:
-            self._set_voice_state("idle", source="hotkey", last_event="dictation_recording_ignored")
-            return
-
-        self._kick_off_transcribe(audio, source="hotkey")
-
-    def _on_device_voice_start(self, device_id: str, source: AudioSource) -> bool:
-        """Begin a device-driven voice-typing session.
-
-        Returns ``False`` (so the WS dispatcher emits a
-        ``session_busy`` error frame) when:
-        - a meeting is active and this device is *not* attached
-          to it (meeting holds the audio floor);
-        - the local hotkey or another device already owns the
-          voice session.
-
-        When this device *is* attached to the active meeting, the
-        meeting already started its recorder — return ``True``
-        without claiming the voice session so the WS keeps pumping
-        binary frames into the meeting's drain path.
-        """
-        active_meeting = self._active_meeting_session()
-        if active_meeting is not None:
-            if active_meeting.is_device_attached(device_id):
-                return True
-            self._set_runtime_activity(
-                "complete",
-                source="device",
-                label="Device busy",
-                detail="Meeting audio is already active.",
-                last_event="device_dictation_busy",
-                last_error="",
-            )
-            return False
-        from .agent_context import get_recent_awaiting_agent_session
-
-        agent_reply_session = get_recent_awaiting_agent_session(max_age_seconds=120)
-        if not self._agent_reply_deliverable(agent_reply_session):
-            with self.state_lock:
-                self.runtime_status["last_error"] = "Agent reply target unavailable"
-            self.device_status.send(device_id, "No reply target", ttl_ms=3000)
-            log.info(
-                "device_voice_start_rejected_no_agent_reply_target",
-                extra={"device_id": device_id},
-            )
-            self._set_runtime_activity(
-                "error",
-                source="device",
-                detail="Agent reply target unavailable.",
-                last_event="device_dictation_no_reply_target",
-                last_error="Agent reply target unavailable",
-            )
-            return False
-        owner = f"device:{device_id}"
-        try:
-            accepted = self.voice_session.begin(source, owner=owner)
-        except Exception as exc:
-            log.error(f"Device voice-typing start failed: {exc}")
-            self._set_runtime_activity(
-                "error",
-                source="device",
-                detail="Device recording failed.",
-                last_event="device_dictation_recording_failed",
-                last_error=f"{type(exc).__name__}: {exc}",
-            )
-            return False
-        if not accepted:
-            self._set_runtime_activity(
-                "complete",
-                source="device",
-                label="Device busy",
-                detail="Another HoldSpeak audio session is active.",
-                last_event="device_dictation_busy",
-                last_error="",
-            )
-            return False
-        self._set_voice_state(
-            "recording",
-            source="device",
-            detail="Device voice input is recording.",
-            last_event="device_dictation_recording_started",
-            last_error="",
-        )
-        # AIPI-4-13: TX state lives in the device's firmware-side
-        # tx_label glyph (top-right, ↑ during right-button hold).
-        # No "Listening..." pushback — would clobber the bottom
-        # widget's persistent meeting/idle text.
-        return True
-
-    def _on_device_voice_stop(
-        self, device_id: str, source: AudioSource
-    ) -> Optional[np.ndarray]:
-        active_meeting = self._active_meeting_session()
-        if active_meeting is not None and active_meeting.is_device_attached(device_id):
-            # The meeting owns this device's audio routing.
-            return None
-        owner = f"device:{device_id}"
-        try:
-            audio = self.voice_session.end(owner=owner)
-        except Exception as exc:
-            log.error(f"Device voice-typing stop failed: {exc}")
-            self._set_voice_state(
-                "idle",
-                source="device",
-                detail="Device recording stop failed.",
-                last_event="device_dictation_recording_stop_failed",
-                last_error=f"{type(exc).__name__}: {exc}",
-            )
-            return None
-        if audio is None:
-            self._set_voice_state("idle", source="device", last_event="device_dictation_recording_ignored")
-            return None
-
-        # AIPI-4-13: no "Thinking..." pushback to bottom — the absent
-        # tx_label arrow after release already signals "we're done
-        # capturing, processing now". Transcript snippet lands in
-        # the middle slot below.
-
-        from .agent_context import get_recent_awaiting_agent_session
-
-        agent_reply_session = get_recent_awaiting_agent_session(max_age_seconds=120)
-
-        def _device_transcript_complete(text: str) -> None:
-            snippet = (text or "").strip()[:150]
-            self.device_status.send(device_id, snippet, ttl_ms=4000)
-
-        self._kick_off_transcribe(
-            audio,
-            on_complete=_device_transcript_complete,
-            agent_reply_session=agent_reply_session,
-            source="device",
-        )
-        return audio
-
-    def _on_device_voice_cancel(self, device_id: str) -> None:
-        self.voice_session.cancel(owner=f"device:{device_id}")
-
-    def _on_device_event(self, device_id: str, name: str, at: Optional[float]) -> None:
-        """Inbound device event from the WS.
-
-        - ``long_press`` (HS-14-07): bookmark gesture in an active
-          meeting where this device is attached.
-        - ``double_left_click`` (AIPI-4-14): cycle through meeting-stat
-          views (Numbers → Speakers → Intel) on the device's middle
-          slot.
-        """
-        if name == "long_press":
-            active = self._active_meeting_session()
-            if active is None or not active.is_device_attached(device_id):
-                return
-            bookmark = active.add_bookmark(label="", auto_label=True)
-            if bookmark is None:
-                return
-            attached_ids = [d.id for d in active.state.devices] if active.state else []
-            self.device_status.broadcast(
-                attached_ids,
-                f"Bookmark @ {bookmark.timestamp:.0f}s",
-                ttl_ms=2500,
-            )
-            return
-
-        if name == "double_left_click":
-            active = self._active_meeting_session()
-            if active is None or active.state is None:
-                log.info(
-                    "device_event_double_tap_ignored",
-                    extra={"device_id": device_id, "reason": "no_active_meeting"},
-                )
-                return
-            if not active.is_device_attached(device_id):
-                log.info(
-                    "device_event_double_tap_ignored",
-                    extra={"device_id": device_id, "reason": "device_not_attached"},
-                )
-                return
-            current = self.device_stats_cycle.get(device_id, -1)
-            next_index, view_id, formatter = pick_next_view(current)
-            self.device_stats_cycle[device_id] = next_index
-            payload = formatter(active.state)
-            self.device_status.send(device_id, payload, ttl_ms=4000)
-            log.info(
-                "device_event_double_tap_cycled",
-                extra={"device_id": device_id, "view": view_id, "index": next_index},
-            )
-            return
-
-        log.info(
-            "device_event_ignored",
-            extra={"device_id": device_id, "event_name": name},
-        )
-
-    def _on_device_health(self, descriptor: object) -> None:
-        active = self._active_meeting_session()
-        if active is None:
-            return
-        updater = getattr(active, "update_device_descriptor", None)
-        if callable(updater):
-            updater(descriptor)
-        if self.server is not None:
-            try:
-                from .meeting_session import _device_descriptor_to_dict
-
-                self.server.broadcast("device_health", _device_descriptor_to_dict(descriptor))
-            except Exception as exc:
-                log.debug(f"Failed to broadcast device health: {exc}")
-
-    def _on_device_query(
-        self,
-        device_id: str,
-        name: str,
-        at: Optional[float],
-    ) -> Optional[dict[str, object]]:
-        from .agent_context import (
-            get_recent_awaiting_agent_session,
-            select_next_awaiting_agent_session,
-        )
-        from .agent_device import (
-            AGENT_NEXT_QUERY,
-            AGENT_QUERY_NAMES,
-            build_agent_query_response,
-        )
-
-        if name in AGENT_QUERY_NAMES:
-            if name == AGENT_NEXT_QUERY:
-                session = select_next_awaiting_agent_session(max_age_seconds=120)
-                response_name = "agent_status"
-            else:
-                session = get_recent_awaiting_agent_session(max_age_seconds=120)
-                response_name = name
-            response = build_agent_query_response(response_name, session)
-            if response is not None:
-                return response
-
-        if name != "last_segment":
-            log.info(
-                "device_query_ignored",
-                extra={"device_id": device_id, "query_name": name, "at": at},
-            )
-            return {"text": f"Unknown query: {name}"[:500], "ttl_ms": 3000}
-        active = self._active_meeting_session()
-        if active is None or active.state is None:
-            return {"text": "No transcript yet", "ttl_ms": 5000}
-        for segment in reversed(active.state.segments):
-            if getattr(segment, "device_id", None) != device_id:
-                continue
-            speaker = getattr(segment, "speaker", None) or "?"
-            text = getattr(segment, "text", "") or ""
-            return {
-                "text": f"{speaker}: {text}"[:500],
-                "ttl_ms": 5000,
-            }
-        return {"text": "No transcript yet", "ttl_ms": 5000}
 
     def _signal_handler(self, sig, frame) -> None:
         _ = sig, frame
