+++ tests/unit/test_wake_runtime.py
@@ -16,6 +16,7 @@ from types import SimpleNamespace
 import numpy as np
 import pytest
 
+import holdspeak.runtime.dictation_capture as dictation_capture
 import holdspeak.web_runtime as web_runtime
 from holdspeak.voice_typing import VoiceTypingSession
 from holdspeak.wake_word import FRAME_SAMPLES, ArmedCapture
@@ -153,7 +154,7 @@ def test_preview_default_never_types(monkeypatch):
         seen.update(kwargs, text=text)
         return text.upper()
 
-    monkeypatch.setattr(web_runtime, "run_dictation_pipeline", fake_pipeline)
+    monkeypatch.setattr(dictation_capture, "run_dictation_pipeline", fake_pipeline)
     rt._transcribe_wake(np.zeros(16000, dtype=np.float32))
 
     assert rt.typer.typed == []  # the condition: preview NEVER types
